+++ src/core_modules/capture-core/components/Pages/EnrollmentEditEvent/EnrollmentEditEventPage.component.js
@@ -7,10 +7,33 @@ import withStyles from '@material-ui/core/styles/withStyles';
 import type { PlainProps } from './EnrollmentEditEventPage.types';
 import { pageMode } from './EnrollmentEditEventPage.const';
 import { WidgetEventEdit } from '../../WidgetEventEdit/';
+import { WidgetError } from '../../WidgetErrorAndWarning/WidgetError';
+import { WidgetWarning } from '../../WidgetErrorAndWarning/WidgetWarning';
+import { WidgetFeedback } from '../../WidgetFeedback';
+import { WidgetIndicator } from '../../WidgetIndicator';
+import { WidgetProfile } from '../../WidgetProfile';
+import { WidgetEnrollment } from '../../WidgetEnrollment';
 
 const styles = ({ typography }) => ({
     page: {
         margin: spacersNum.dp16,
+    },
+    columns: {
+        display: 'flex',
+    },
+    leftColumn: {
+        flexGrow: 3,
+        flexShrink: 1,
+        width: 872,
+    },
+    rightColumn: {
+        flexGrow: 1,
+        flexShrink: 1,
+        paddingLeft: spacersNum.dp16,
+        width: 360,
+        display: 'flex',
+        flexDirection: 'column',
+        gap: '12px',
     },
     title: {
         ...typography.title,
@@ -21,7 +44,14 @@ const styles = ({ typography }) => ({
 const EnrollmentEditEventPagePain = ({
     mode,
     programStage,
+    teiId,
+    enrollmentId,
+    programId,
+    widgetEffects,
+    hideWidgets,
+    onDelete,
     classes,
+    onGoBack,
 }: PlainProps) => (
     <div className={classes.page}>
         <div className={classes.title}>
@@ -33,11 +63,41 @@ const EnrollmentEditEventPagePain = ({
                     escape: ':',
                 })}
         </div>
-        {programStage ? (
-            <WidgetEventEdit programStage={programStage} />
-        ) : (
-            <span>{i18n.t('We could not find the stage in the program')}</span>
-        )}
+        <div className={classes.columns}>
+            <div className={classes.leftColumn}>
+                {programStage ? (
+                    <WidgetEventEdit programStage={programStage} onGoBack={onGoBack} />
+                ) : (
+                    <span>{i18n.t('We could not find the stage in the program')}</span>
+                )}
+            </div>
+            <div className={classes.rightColumn}>
+                <WidgetError error={widgetEffects.errors} />
+                <WidgetWarning warning={widgetEffects.warnings} />
+                {!hideWidgets.feedback && (
+                    <WidgetFeedback
+                        emptyText={i18n.t('There are no feedback for this event')}
+                        feedback={widgetEffects.feedbacks}
+                    />
+                )}
+                {!hideWidgets.indicator && (
+                    <WidgetIndicator
+                        emptyText={i18n.t('There are no indicators for this event')}
+                        indicators={widgetEffects.indicators}
+                    />
+                )}
+                <WidgetProfile
+                    teiId={teiId}
+                    programId={programId}
+                />
+                <WidgetEnrollment
+                    teiId={teiId}
+                    enrollmentId={enrollmentId}
+                    programId={programId}
+                    onDelete={onDelete}
+                />
+            </div>
+        </div>
     </div>
 );