+++ src/core_modules/capture-core/components/Pages/EnrollmentEditEvent/EnrollmentEditEventPage.types.js
@@ -1,8 +1,16 @@
 // @flow
 import type { ProgramStage } from '../../../metaData';
+import type { WidgetEffects, HideWidgets } from '../Enrollment/EnrollmentPageDefault/EnrollmentPageDefault.types';
 
 export type PlainProps = {|
     programStage: ?ProgramStage,
+    widgetEffects: WidgetEffects,
+    hideWidgets: HideWidgets,
+    teiId: string,
+    enrollmentId: string,
+    programId: string,
     mode: string,
+    onDelete: () => void,
+    onGoBack: () => void,
     ...CssClasses,
 |};