+++ src/core_modules/capture-core/components/ScopeSelector/hooks/useReset.js
@@ -1,0 +1,12 @@
+// @flow
+import { useHistory } from 'react-router';
+
+export const useReset = () => {
+    const history = useHistory();
+
+    const reset = () => {
+        history.push('/');
+    };
+
+    return { reset };
+};