+++ src/core_modules/capture-core/components/WidgetStagesAndEvents/Stages/Stage/StageDetail/stageDetail.types.js
@@ -1,0 +1,12 @@
+// @flow
+import type { ApiTEIEvent } from 'capture-core/events/getEnrollmentEvents';
+import type { StageDataElement } from '../../../types/common.types';
+
+export type Props = {|
+    events: Array<ApiTEIEvent>,
+    dataElements: Array<StageDataElement>,
+    eventName: string,
+    onEventClick: (eventId: string, stageId: string) => void,
+    ...CssClasses,
+|};
+