+++ src/core_modules/capture-core/components/WidgetStagesAndEvents/Stages/Stage/stage.types.js
@@ -1,8 +1,11 @@
 // @flow
-import type { ProgramStage } from '../../../../metaData';
+import type { ApiTEIEvent } from 'capture-core/events/getEnrollmentEvents';
+import type { Stage } from '../../types/common.types';
 
 export type Props = {|
-    stage: ProgramStage,
+    stage: Stage,
+    events: Array<ApiTEIEvent>,
     className?: string,
+    onEventClick: (eventId: string, stageId: string) => void,
     ...CssClasses,
 |};