+++ src/core_modules/capture-core/components/WidgetStagesAndEvents/Stages/Stages.component.js
@@ -1,24 +1,21 @@
 // @flow
 import React, { type ComponentType } from 'react';
 import { withStyles } from '@material-ui/core';
-import { spacersNum } from '@dhis2/ui';
 import { Stage } from './Stage';
 import type { Props } from './stages.types';
 
-const styles = {
-    stage: {
-        marginBottom: spacersNum.dp16,
-    },
-};
-export const StagesPlain = ({ stages, classes }: Props) => (
+const styles = {};
+export const StagesPlain = ({ stages, events, classes, onEventClick }: Props) => (
     <>
         {
-            [...stages.values()]
+            stages
                 .map(stage => (
                     <Stage
+                        events={events?.filter(event => event.programStage === stage.id)}
                         key={stage.id}
                         stage={stage}
                         className={classes.stage}
+                        onEventClick={onEventClick}
                     />
                 ))
         }