+++ src/core_modules/capture-core/components/WidgetStagesAndEvents/stagesAndEvents.types.js
@@ -1,7 +1,10 @@
 // @flow
-import type { ProgramStage } from '../../metaData';
+import type { ApiTEIEvent } from 'capture-core/events/getEnrollmentEvents';
+import type { Stage } from './types/common.types';
 
 export type Props = {|
-    stages: Map<string, ProgramStage>,
+    stages: Array<Stage>,
+    events: Array<ApiTEIEvent>,
+    onEventClick: (eventId: string, stageId: string) => void,
     className?: string,
 |};